Refactor Day03 parts to shared run helper
Extract common parsing logic into run(path, indexSize) and have
RunPartOne and RunPartTwo delegate to it. Change RunPartOne return type
to int64 and update tests to expect int64

+++ 2025/day03/main_test.go
@@ -4,7 +4,7 @@ import "testing"
 
 func TestPartOne(t *testing.T) {
 	got := RunPartOne("test_input")
-	want := 357
+	want := int64(357)
 
 	if got != want {
 		t.Errorf("got %d want %d", got, want)
